requirements.txt
python-telegram-bot
psutil
python-dotenv